requirements_base.txt
# Pinning setuptools version to 59.6.0 due to CircleCI failure with 60.0.4
setuptools==59.6.0

alabaster==0.7.12         # via sphinx
alembic==1.4.2            # via -r requirements.in
aniso8601==8.0.0          # via flask-restful
astroid==2.11.0            # via pylint
babel==2.9.1              # via sphinx
backoff==1.10.0           # via -r requirements.in
blinker==1.4              # via -r requirements.in
cachetools==4.1.0         # via google-auth
certifi==2022.12.7       # via geventhttpclient, requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
click==8.0.0              # via flask, pip-tools
configargparse==1.2.3     # via locust
coverage==6.0b1             # via -r requirements.in
cryptography==36.0.2       # via oauthlib, pyopenssl, requests
defusedxml==0.6.0         # via jira
dictalchemy3==1.0.0       # via -r requirements.in
dnspython==1.16.0         # via -r requirements.in
docutils==0.16            # via sphinx
faker==4.1.0              # via -r requirements.in
fhirclient==3.2.0         # via -r requirements.in
flask-basicauth==0.2.0    # via locust
flask-restful==0.3.9      # via -r requirements.in
flask==2.0.2              # via -r requirements.in, flask-basicauth, flask-restful, locust
flask-limiter==1.4
gevent==21.12.0            # via geventhttpclient, locust
geventhttpclient==2.0.2   # via locust
google-api-core[grpc]==1.17.0  # via google-api-python-client, google-cloud-bigquery, google-cloud-core, google-cloud-datastore, google-cloud-firestore, google-cloud-logging, google-cloud-tasks
google-api-python-client==1.8.3  # via -r requirements.in, google-python-cloud-debugger
google-auth-httplib2==0.0.3  # via google-api-python-client, google-python-cloud-debugger
google-auth==1.28.0       # via google-api-core, google-api-python-client, google-auth-httplib2, google-cloud-bigquery, google-cloud-storage, google-python-cloud-debugger
google-cloud-bigquery==1.24.0  # via -r requirements.in
google-cloud-core==1.3.0  # via google-cloud-bigquery, google-cloud-datastore, google-cloud-firestore, google-cloud-logging, google-cloud-storage
google-cloud-datastore==1.12.0  # via -r requirements.in
google-cloud-firestore==1.7.0  # via -r requirements.in
google-cloud-logging==1.15.0  # via -r requirements.in
google-cloud-storage==1.28.1  # via -r requirements.in
google-cloud-tasks==1.5.0  # via -r requirements.in
google-resumable-media==0.5.0  # via google-cloud-bigquery, google-cloud-storage
googleapis-common-protos[grpc]==1.51.0  # via google-api-core, grpc-google-iam-v1
googlemaps==4.4.1         # via -r requirements.in
greenlet==1.1.3          # via gevent
grpc-google-iam-v1==0.12.3  # via google-cloud-tasks
grpcio==1.41.1           # via google-api-core, googleapis-common-protos, grpc-google-iam-v1
gspread==3.7.0
gunicorn==20.0.4          # via -r requirements.in
httplib2==0.19.1          # via google-api-python-client, google-auth-httplib2, oauth2client
idna==2.9                 # via requests
imagesize==1.2.0          # via sphinx
isodate==0.6.0            # via fhirclient
isort==4.3.21             # via pylint
itsdangerous==2.0.1       # via flask
jinja2==3.0.2             # via flask, sphinx
jira==2.0.0               # via -r requirements.in
lazy-object-proxy==1.4.3  # via astroid
locust==1.3.2             # via -r requirements.in
mako==1.2.2               # via alembic
markupsafe==2.0.1         # via jinja2, mako
marshmallow-jsonschema==0.10.0  # via -r requirements.in
marshmallow==3.6.0        # via -r requirements.in, marshmallow-jsonschema
mccabe==0.6.1             # via pylint
mock==4.0.3               # via -r requirements.in
msgpack==1.0.0            # via locust
netaddr==0.7.19           # via -r requirements.in
oauth2client==4.1.3       # via -r requirements.in
oauthlib[signedtoken]==3.1.0  # via jira, requests-oauthlib
packaging==20.4           # via sphinx
#pandas==1.3.1
parameterized==0.7.4      # via -r requirements.in
pbr==5.4.5                # via jira
pdfminer.six==20201018
# pip-tools==5.1.2          # via -r requirements.in
protobuf==3.20.3          # via -r requirements.in, google-api-core, google-cloud-bigquery, googleapis-common-protos
protorpc==0.12.0          # via -r requirements.in
psutil==5.7.0             # via locust
psycopg2-binary==2.8.6
pyasn1-modules==0.2.8     # via google-auth, oauth2client
pyasn1==0.4.8             # via oauth2client, pyasn1-modules, rsa
pycparser==2.20           # via cffi
pygments==2.7.4           # via sphinx
pyjwt==2.4.0              # via oauthlib
pylint==2.13.0            # via -r requirements.in
pyopenssl==19.1.0         # via requests
pyparsing==2.4.7          # via packaging
python-dateutil==2.8.1    # via alembic, faker
python-editor==1.0.4      # via alembic
python-http-client==3.2.7  # via sendgrid
pytz==2020.1              # via babel, flask-restful, google-api-core, google-cloud-firestore
pyyaml==5.4.1             # via google-python-cloud-debugger
pyzmq==19.0.1             # via locust
redis==3.5.3              # via flask-limiter
requests-oauthlib==1.3.0  # via jira
requests-toolbelt==0.9.1  # via jira
requests[security]==2.25.1  # via -r requirements.in, fhirclient, google-api-core, googlemaps, jira, locust, requests-oauthlib, requests-toolbelt, sphinx
rsa==4.7.2                  # via google-auth, oauth2client
sendgrid==6.9.1           # via -r requirements.in
simple-geometry==0.1.4  # geometry
simplejson==3.17.0        # via -r requirements.in
six==1.15.0               # via astroid, cryptography, flask-restful, geventhttpclient, google-api-core, google-api-python-client, google-auth, google-cloud-bigquery, google-python-cloud-debugger, google-resumable-media, grpcio, isodate, jira, oauth2client, packaging, pip-tools, protobuf, protorpc, pyopenssl, python-dateutil
snowballstemmer==2.0.0    # via sphinx
sphinx==3.0.4             # via -r requirements.in
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
sqlalchemy==1.3.17        # via -r requirements.in, alembic, dictalchemy
sqlparse==0.4.2           # via -r requirements.in
supervisor==4.2.0         # via -r requirements.in
text-unidecode==1.3       # via faker
toml==0.10.1              # via pylint
typed-ast==1.4.1          # via astroid
uritemplate==3.0.1        # via google-api-python-client
urllib3==1.26.5           # via requests
werkzeug==2.0.2           # via flask
wrapt==1.12.1             # via astroid
xmltodict==0.12.0         # via -r requirements.in
zope.event==4.4           # via gevent
zope.interface==5.1.0     # via gevent
# DA-3099 -- dchan
Flask-GraphQL==2.0.0
graphene==3.1.0
graphql-core==3.1
names
